requirements.txt
requests
python-engineio>=v3.7.0
python-socketio
